Working file based build
a field-book is a folder, containing src/ for all cells as .ojs files
named,unnamed,imports are subfolders in src to store different types of cells, on save/edit/delete the nodejs backend watchign files in the src folder will send events to front end which has the compiler and a runtime, which injects the changes and maintains cell dom nodes in place. The code automatically launches a webpage, once user is connected, the filewatchers are started for each connected client, this is inefficient but works very well for local editing for single user. The client first looks for an array of files to get the order as there is no implicit order from files in 3 folders.  if not, it tries to build the array as cells are added, the array can then be stored back to file (TBD), there is also a slight customization to the ui as this will not have inline code editor.